ui_dashboard/requirements.txt
streamlit==1.29.0   # Pinned for consistency
psycopg2-binary>=2.9,<3.0
python-dotenv>=1.0,<2.0 # Au cas où
pandas==2.1.4       # Pinned for consistency
